refactor SidiBarahniPreRoundActivity
